fault_tolerence/requirements.txt
Flask==2.3.0
requests==2.31.0
grpcio
grpcio-tools